refactor(component):restruct code for seprate container and component
