--- src/components/TravelSpots/SpotContainer.js
@@ -1,14 +1,0 @@
-
-
-import TravelSpots from "./TravelSpots";
-
-
-export default function SpotContainer() {
-
-    return (
-        <div>
-            <TravelSpots />
-        </div>
-    );
-}
-
