+++ src/components/TravelSpots/travelspots.module.css
@@ -1,4 +1,10 @@
 /* HotSpot part*/
+.SpotContainer {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
 
 .hotSpotGroup {
